refactor: rename app to internal directory
