+++ internal/domain/diary.go
