+++ main.go
@@ -1,19 +1,19 @@
-package main
+// package main
 
-import (
-	"net/http"
+// import (
+// 	"net/http"
 
-	"github.com/labstack/echo/v4"
-)
+// 	"github.com/labstack/echo/v4"
+// )
 
-func main() {
-	e := echo.New()
+// func main() {
+// 	e := echo.New()
 
-	// Health check
-	e.GET("/health", func(c echo.Context) error {
-		return c.String(http.StatusOK, "OK from Go backend!")
-	})
+// 	// Health check
+// 	e.GET("/health", func(c echo.Context) error {
+// 		return c.String(http.StatusOK, "OK from Go backend!")
+// 	})
 
-	// サーバ起動
-	e.Logger.Fatal(e.Start(":8080"))
-}
+// 	// サーバ起動
+// 	e.Logger.Fatal(e.Start(":8080"))
+// }
